Added a dto for product updates
I wanted to keep the variable 'active' as a boolean since it's not nullable in the database, but ran into the issue that I didn't have a value to represent no change. After a lot of reading and thinking about it - this was the only solution that made sense to me.